+++ src/app/layout.js
@@ -6,8 +6,8 @@ import "@fontsource/roboto/300.css";
 import "@fontsource/roboto/400.css";
 import "@fontsource/roboto/500.css";
 import "@fontsource/roboto/700.css";
-import Navbar from "@/app/navbar/page";
-import { GlobalContextProvider } from "./context";
+import Navbar from "@/app/components/navbar";
+import { GlobalContextProvider } from "./components/context";
 
 import FlashMsg from "./components/flashMsg";
